+++ weather-service/src/main/java/com/waither/weatherservice/service/WeatherService.java
@@ -1,11 +1,11 @@
 package com.waither.weatherservice.service;
 
+import static com.waither.weatherservice.utills.DateTimeUtils.*;
+
 import java.io.IOException;
 import java.net.URISyntaxException;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.Arrays;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
@@ -18,12 +18,12 @@ import com.waither.weatherservice.entity.ExpectedWeather;
 import com.waither.weatherservice.entity.Region;
 import com.waither.weatherservice.entity.WeatherAdvisory;
 import com.waither.weatherservice.exception.WeatherExceptionHandler;
-import com.waither.weatherservice.gps.GpsTransfer;
+import com.waither.weatherservice.utills.GpsTransferUtils;
 import com.waither.weatherservice.kafka.KafkaMessage;
 import com.waither.weatherservice.kafka.Producer;
 import com.waither.weatherservice.openapi.ForeCastOpenApiResponse;
 import com.waither.weatherservice.openapi.MsgOpenApiResponse;
-import com.waither.weatherservice.openapi.OpenApiUtil;
+import com.waither.weatherservice.openapi.OpenApiUtils;
 import com.waither.weatherservice.repository.DailyWeatherRepository;
 import com.waither.weatherservice.repository.ExpectedWeatherRepository;
 import com.waither.weatherservice.repository.RegionRepository;
@@ -39,7 +39,7 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class WeatherService {
 
-	private final OpenApiUtil openApiUtil;
+	private final OpenApiUtils openApiUtil;
 	private final DailyWeatherRepository dailyWeatherRepository;
 	private final ExpectedWeatherRepository expectedWeatherRepository;
 	private final WeatherAdvisoryRepository weatherAdvisoryRepository;
@@ -134,9 +134,9 @@ public class WeatherService {
 
 	public void createWeatherAdvisory(double latitude, double longitude) throws URISyntaxException, IOException {
 		LocalDate now = LocalDate.now();
-		String today = openApiUtil.convertLocalDateToString(now);
+		String today = convertLocalDateToString(now);
 
-		String location = GpsTransfer.convertGpsToRegionCode(latitude, longitude);
+		String location = GpsTransferUtils.convertGpsToRegionCode(latitude, longitude);
 
 		List<MsgOpenApiResponse.Item> items = openApiUtil.callAdvisoryApi(location, today);
 
@@ -205,38 +205,6 @@ public class WeatherService {
 		log.info(regionName + "[Main - api] MainWeatherResponse : {}", weatherMainResponse);
 
 		return weatherMainResponse;
-	}
-
-	public String convertLocalDateTimeToString(LocalDateTime time) {
-		DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
-		String formattedDateTime = time.format(formatter);
-
-		String[] lst = formattedDateTime.split(" ");
-		String baseDate = lst[0].replace("-", "");
-
-		String[] temp = lst[1].split(":");
-		String baseTime = temp[0] + "00";
-
-		return baseDate + "_" + baseTime;
-	}
-
-	public LocalDateTime convertLocalDateTimeToDailyWeatherTime(LocalDateTime time) {
-
-		// DailyWeather 정보는 3시간마다
-		List<Integer> scheduledHours = Arrays.asList(0, 3, 6, 9, 12, 15, 18, 21);
-
-		int currentHour = time.getHour();
-		int adjustedHour = scheduledHours.stream()
-			.filter(hour -> hour <= currentHour)
-			.reduce((first, second) -> second)
-			.orElse(scheduledHours.get(scheduledHours.size() - 1)); // 이전 날의 마지막 스케줄 시간(21시) 반환
-
-		// 현재 시간이 첫 스케줄 시간(0시)보다 작을 경우, 전날의 마지막 스케줄 시간으로 설정
-		if (currentHour < scheduledHours.get(0)) {
-			time = time.minusDays(1);
-		}
-
-		return time.withHour(adjustedHour).withMinute(0).withSecond(0).withNano(0);
 	}
 
 	public List<Region> getRegionList() {
